fix signedness of comparison

+++ src/gf256.c
@@ -83,7 +83,7 @@ uint8_t gf256_add(uint8_t a, uint8_t b)
 
 uint8_t gf256_exp(int e)
 {
-	assert(e < sizeof(OCT_EXP));
+	assert(e < (int)sizeof(OCT_EXP));
 	return OCT_EXP[e];
 }
